tooling: webc registry publishing, unified publication rels
