+++ .github/workflows/publish-index.yml
@@ -66,7 +66,7 @@ jobs:
         # Until then, don't fail on publish errors, so that the GH Pages deployment still fires.
         continue-on-error: true
         run: |
-          nix run .#scripts.publish -- \
+          nix run .#scripts.publish-index -- \
             --registry "$(readlink -f registry-result)" \
             --rev "${{ github.event.workflow_run.head_sha }}"
