--- pkgs/python-registry/rels.json
@@ -1,11 +1,0 @@
-{
-  "tiktoken": {
-    "0.12.0": 3
-  },
-  "tokenizers": {
-    "0.22.2": 3
-  },
-  "wrapt": {
-    "1.17.2": 2
-  }
-}
