+++ rels.json
@@ -1,0 +1,11 @@
+{
+  "pythonRegistry.wheels.tiktoken": {
+    "0.12.0": 3
+  },
+  "pythonRegistry.wheels.tokenizers": {
+    "0.22.2": 3
+  },
+  "pythonRegistry.wheels.wrapt": {
+    "1.17.2": 2
+  }
+}
